refactor: insertDB関数をdao_testに移動
