+++ app/dao/relationship_test.go
@@ -2,37 +2,11 @@ package dao_test
 
 import (
 	"context"
-	"strconv"
 	"testing"
 	"yatter-backend-go/app/domain/object"
 
 	"github.com/stretchr/testify/assert"
 )
-
-func insertAccountDB(t *testing.T, ctx context.Context, accounts []object.Account) {
-	for _, account := range accounts {
-		err := accountRepo.Create(ctx, &account)
-		assert.NoError(t, err)
-	}
-}
-
-func insertRelationshipDB(t *testing.T, ctx context.Context, relationships []object.Relationship) {
-	for _, relationship := range relationships {
-		err := relationshipRepo.Create(ctx, relationship.FollowingId, relationship.FollowerId)
-		assert.NoError(t, err)
-	}
-}
-
-func createAccountObject(num int) []object.Account {
-	accounts := make([]object.Account, num)
-	for i := 0; i < num; i++ {
-		accounts[i] = object.Account{
-			Username:     "test" + strconv.Itoa(i),
-			PasswordHash: "password",
-		}
-	}
-	return accounts
-}
 
 func TestRelationshipCreate(t *testing.T) {
 	ctx := context.Background()
